Refactor deployment script
